move options out of urlOpts.ts
