+++ packages/expo-cli/src/urlOpts.ts
@@ -21,27 +21,6 @@ function addOptions(program: Command) {
     .option('--tunnel', 'Same as --host tunnel')
     .option('--lan', 'Same as --host lan')
     .option('--localhost', 'Same as --host localhost');
-  if (program._name.includes('start')) {
-    program
-      .option('--dev', 'Turns dev flag on')
-      .option('--no-dev', 'Turns dev flag off')
-      .option('--minify', 'Turns minify flag on')
-      .option('--no-minify', 'Turns minify flag off')
-      .option('--https', 'To start webpack with https protocol')
-      .option('--no-https', 'To start webpack with http protocol');
-  }
-}
-
-function hasBooleanArg(rawArgs: string[], argName: string) {
-  return rawArgs.includes('--' + argName) || rawArgs.includes('--no-' + argName);
-}
-
-function getBooleanArg(rawArgs: string[], argName: string) {
-  if (rawArgs.includes('--' + argName)) {
-    return true;
-  } else {
-    return false;
-  }
 }
 
 async function optsAsync(projectDir: string, options: any) {
@@ -64,21 +43,6 @@ async function optsAsync(projectDir: string, options: any) {
     opts.hostType = 'localhost';
   } else {
     opts.hostType = 'lan';
-  }
-
-  let rawArgs = options.parent.rawArgs;
-
-  if (hasBooleanArg(rawArgs, 'dev')) {
-    opts.dev = getBooleanArg(rawArgs, 'dev');
-  } else {
-    opts.dev = true;
-    opts.minify = false;
-  }
-  if (hasBooleanArg(rawArgs, 'minify')) {
-    opts.minify = getBooleanArg(rawArgs, 'minify');
-  }
-  if (hasBooleanArg(rawArgs, 'https')) {
-    opts.https = getBooleanArg(rawArgs, 'https');
   }
 
   await ProjectSettings.setAsync(projectDir, opts);
